clear modules
